# openapi.yaml
openapi: 3.1.0
info:
  title: Browser Logger System API
  description: |
    A comprehensive HTTP API for collecting, storing, and retrieving browser logs in real-time.

    The Browser Logger System provides a lightweight infrastructure for capturing frontend console logs and application events, making them accessible via HTTP endpoints and Server-Sent Events (SSE).

    ## Architecture Overview

    The system consists of three main components:

    1. **Frontend Logger** (`mcp-logger.js`) - Captures browser console logs and application events
    2. **Backend Server** (this API) - HTTP server with in-memory log storage
    3. **MCP Server** - Integrates with AI assistants via Model Context Protocol

    ## Key Features

    - **Real-time log collection** via HTTP POST endpoints
    - **Multi-application support** with isolated log storage
    - **Multi-host tracking** for different frontend environments
    - **Namespace isolation** for organizing log types
    - **Server-Sent Events (SSE)** for real-time log streaming
    - **Duplicate filtering** to prevent log spam
    - **Rate limiting** for performance protection
    - **In-memory storage** with configurable limits

    ## Quick Start

    1. **Start the server**: The API runs on `http://localhost:22345` by default
    2. **Submit logs**: POST to `/api/logs/submit` with your log data
    3. **Retrieve logs**: GET from `/api/logs/:app/:host/:namespace`
...

    ## Data Model

    ### Applications
...

    ### Hosts
...

...

...
servers:
...
paths:
...
  /api/logs/submit:
    post:
      tags:
        - Logs
      summary: Submit log entries
      description: |
        Submit one or more log entries to be stored in the system.

        This is the primary endpoint for frontend applications to send logs to the backend server.

        ## Request Format

        The request body must contain:
        - `app`: Application identifier (required)
        - `host`: Frontend host origin (required)
        - `logs`: Object with namespace keys and log data (required)

        ## Log Namespaces

        ### Browser Logs (`browser`)
        Array of console log entries with the following structure:
        ```json
        {
          "level": "INFO|ERROR|WARN|LOG|DEBUG",
          "message": "Log message",
          "source": "file.js:line",
          "timestamp": 1760710400000
        }
        ```

        ### Application Logs (custom namespaces)
        Custom data objects for application-specific events:
        ```json
        {
          "action": "button_click",
          "element": "submit-btn",
          "metadata": { "key": "value" }
        }
        ```

        ## Rate Limiting

        - **Localhost**: 200 requests per 10 seconds
        - **Remote**: 1000 requests per minute
        - Duplicate filtering is applied to prevent log spam

        ## Response Types

        - `success`: Logs were stored successfully
        - `skipped`: All logs were filtered as duplicates
      operationId: submitLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogSubmissionRequest'
            examples:
              browser_logs:
                summary: Browser console logs
                value:
                  app: my-frontend-app
                  host: localhost:3000
                  logs:
                    browser:
                      - level: INFO
                        message: User logged in successfully
                        source: auth.js:45
                        timestamp: 1760710400000
                      - level: ERROR
                        message: Failed to load user data
                        source: api.js:123
                        timestamp: 1760710401000
              mixed_logs:
                summary: Mixed browser and application logs
                value:
                  app: my-frontend-app
                  host: localhost:3000
                  logs:
                    browser:
                      - level: INFO
                        message: Application initialized
                        source: app.js:10
                        timestamp: 1760710400000
                    user-actions:
                      action: page_view
                      page: /dashboard
                      timestamp: 1760710400000
                    api-calls:
                      endpoint: /api/users
                      method: GET
                      status: 200
                      duration: 150
                      timestamp: 1760710400000
      responses:
        '200':
          description: Logs processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogSubmissionResponse'
              examples:
                success:
                  summary: Logs stored successfully
                  value:
                    status: success
                    stored: 3
                    filtered: 0
                    app: my-frontend-app
                    host: localhost:3000
                    timestamp: 1760710406678
                skipped:
                  summary: All logs filtered as duplicates
                  value:
                    status: skipped
                    reason: All logs filtered as duplicates or spam
                    app: my-frontend-app
                    host: localhost:3000
                    timestamp: 1760710406678
        '400':
          description: Invalid request format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missing_fields:
                  summary: Missing required fields
                  value:
                    error: "Missing required fields: app, host, logs"
                invalid_logs:
                  summary: Invalid logs format
                  value:
                    error: "logs must be an object with namespace keys"
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
              examples:
                rate_limited:
                  summary: Rate limit exceeded
                  value:
                    error: Too many requests
                    retryAfter: 30
                    requestCount: 201
                    limit: 200
...
  /api/logs/stream:
    get:
      tags:
        - Streaming
      summary: Real-time log streaming via Server-Sent Events
      description: |
        Establish a Server-Sent Events (SSE) connection for real-time log streaming.

        This endpoint provides a persistent connection for receiving live log updates as they are submitted to the system.

        ## Connection Flow

        1. Client connects to the SSE endpoint
        2. Server sends a `connected` event
        3. If initial parameters are provided, server sends `initial_logs` event
        4. Server sends `new_logs` events for new log submissions
        5. Server sends `keepalive` events every 30 seconds

        ## Event Types

        ### connected
        Sent when the client first connects:
        ```
        event: connected
        data: {"timestamp": 1760710413076, "message": "Connected to log stream"}
        ```

        ### initial_logs
        Sent if initial parameters are provided:
        ```
        event: initial_logs
        data: {"app": "my-app", "host": "localhost:3000", "namespace": "browser", "logs": [...]}
        ```

        ### new_logs
        Sent when new logs are submitted:
        ```
        event: new_logs
        data: {"host": "localhost:3000", "logs": {"browser": [...], "user-actions": [...]}, "timestamp": 1760710415000}
        ```

        ### keepalive
        Sent every 30 seconds to maintain the connection:
        ```
        event: keepalive
        data: {"timestamp": 1760710433076}
        ```

        ## Usage Examples

        ### JavaScript Client
        ```javascript
        const eventSource = new EventSource('http://localhost:22345/api/logs/stream');

        eventSource.addEventListener('new_logs', (event) => {
          const data = JSON.parse(event.data);
          console.log('New logs:', data.logs);
        });
        ```

        ### curl Command
        ```bash
        curl -N "http://localhost:22345/api/logs/stream?app=my-app&frontend_host=localhost:3000&namespace=browser"
        ```
      operationId: streamLogs
      parameters:
        - name: app
          in: query
          description: Application identifier for initial logs
          schema:
            type: string
          examples:
            app_id:
              summary: Application identifier
              value: my-frontend-app
        - name: frontend_host
          in: query
          description: Frontend host for initial logs
          schema:
            type: string
          examples:
            localhost:
              summary: Local development server
              value: localhost:3000
        - name: namespace
          in: query
          description: Namespace for initial logs
          schema:
            type: string
          examples:
            browser:
              summary: Browser console logs
              value: browser
        - name: lines
          in: query
          description: "Number of initial log entries to send (max: 100)"
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          examples:
            small:
              summary: Small number of initial entries
              value: 10
        - name: filter
          in: query
          description: Text filter for initial log content
          schema:
            type: string
          examples:
            error_filter:
              summary: Filter for error messages
              value: error
      responses:
        '200':
          description: SSE connection established
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
              examples:
                connection:
                  summary: Initial connection events
                  value: |
                    event: connected
                    data: {"timestamp":1760710413076,"message":"Connected to log stream"}

                    event: initial_logs
                    data: {"app":"my-app","host":"localhost:3000","namespace":"browser","logs":[...]}
...
components:
  schemas:
...
    LogSubmissionRequest:
      type: object
      properties:
        app:
          type: string
          description: Application identifier
          example: my-frontend-app
        host:
          type: string
          description: Frontend host origin
          example: localhost:3000
        logs:
          type: object
          description: Log data organized by namespace
          additionalProperties:
            oneOf:
              - type: array
                description: Browser console logs
                items:
                  $ref: '#/components/schemas/BrowserLogEntry'
              - type: object
                description: Application-specific log data
                additionalProperties: true
          example:
            browser:
              - level: INFO
                message: User logged in
                source: auth.js:45
                timestamp: 1760710400000
            user-actions:
              action: button_click
              element: submit-btn
              timestamp: 1760710400000
      required:
        - app
        - host
        - logs
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error description
      required:
        - error
      example:
        error: "Missing required fields: app, host, logs"
        message: The request body must include app, host, and logs fields
...